+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed 
+
+- Don't play sounds from hidden instances
+
 ## [1.0.2] 2025-11-10
 ## [1.0.1] 2025-11-04